param tweaking + bug fixes

+++ rl_baselines/hyperband.py
@@ -3,6 +3,7 @@ import subprocess
 import os
 import shutil
 import glob
+import pprint
 
 from fluentopt.hyperband import hyperband
 import pandas as pd
@@ -23,13 +24,15 @@ def main():
     parser.add_argument('--seed', type=int, default=0, help='random seed (default: 0)')
     parser.add_argument('--srl-model', type=str, default='raw_pixels', choices=list(registered_srl.keys()),
                         help='SRL model to use')
-    parser.add_argument('--num-timesteps', type=int, default=1e4, help='number of timesteps the baseline should run')
+    parser.add_argument('--num-timesteps', type=int, default=2e5, help='number of timesteps the baseline should run')
     parser.add_argument('-v', '--verbose', action='store_true', default=False, help='Display baseline STDOUT')
+    parser.add_argument('--max_iter', type=int, default=100, help='Number of iteration to try')
 
     args, train_args = parser.parse_known_args()
 
     train_args.extend(['--srl-model', args.srl_model, '--seed', str(args.seed), '--algo', args.algo, '--env', args.env,
-                       '--log-dir', "logs/_hyperband_search/", '--num-timesteps', args.num_timesteps, '--no-vis'])
+                       '--log-dir', "logs/_hyperband_search/", '--num-timesteps', str(int(args.num_timesteps)),
+                       '--no-vis'])
 
     if args.verbose:
         # None here means stdout of terminal for subprocess.call
@@ -56,15 +59,22 @@ def main():
         return params
 
     def run_batch(batch):
-        for num_iters, params in batch:
-            printGreen("\nIteration_num={}, Param={}".format(num_iters, params))
+        for i, (num_iters, params) in enumerate(batch):
+            printGreen("\nIteration_num={}, Param:".format(i))
+            pprint.pprint(params)
+            print()
 
-            shutil.rmtree("logs/_hyperband_search/")
+            # cleanup old files
+            if os.path.exists("logs/_hyperband_search/"):
+                shutil.rmtree("logs/_hyperband_search/")
+
+            loop_args = []
 
             # redefine the parsed args for rl_baselines.train
-            loop_args = []
-            for param_name, param_val in params:
-                loop_args.extend([param_name, str(param_val)])
+            if len(params) > 0:
+                loop_args.append("--hyperparam")
+                for param_name, param_val in params.items():
+                    loop_args.append("{}:{}".format(param_name, param_val))
 
             ok = subprocess.call(['python', '-m', 'rl_baselines.train'] + train_args + loop_args, stdout=stdout)
 
@@ -76,7 +86,7 @@ def main():
             assert len(folders) != 0, "Error: Could not find generated directory, halting hyperband search."
             rewards = []
             for montior_path in glob.glob(folders[0] + "/*.monitor.csv"):
-                rewards.append(np.mean(pd.read_csv(montior_path, skiprows=1)["r"]))
+                rewards.append(np.mean(pd.read_csv(montior_path, skiprows=1)["r"][-10:]))
 
             yield -np.mean(rewards)
